COM-3011: move component
