+++ src/components/ImagePickerManagerModal/styles.ts
